hackish attempt to use theano.tile to collapse latent space
The idea was to do a quick hack to parameter tie the latent
space so that each iteration had the same set of latent variables
instead of a new batch. This effectively would collapse the
actual latent space size from z_dim x num_iters to just z_dim
as god intended.

This version half works, which is to say it doesn't really
hold together. It does tie together the parameters when
randomly sampling, but this doesn't seem to prevent the
reconstruction from not sticking to this constraint.

So why am I memorializing this failed experiment with
such fanfare? Because along the way I stumbled upon great
results reconstructing mnist and bmnist. This was either
due to lucky hyperparmater settings (accidently ran without
attention but with iterations), or perhaps my code is
having some sort of strange regularizing effect. Thus
I am committing it as-is, and then will work backward
to reproduce.